Adapt erb templates to match spec + extract common functions in src/pcap/cmd
