+++ jobs/pcap-agent/templates/pcap-agent.yml.erb
@@ -5,6 +5,7 @@ config = {
   "container_store" => p("pcap-agent.container_store"),
   "runc" => p("pcap-agent.runc"),
   "runc_root" => p("pcap-agent.runc_root"),
+  "id" => p("pcap-agent.id")
   "cert" => "/var/vcap/jobs/pcap-agent/config/certs/pcap_agent.crt",
   "key" => "/var/vcap/jobs/pcap-agent/config/certs/pcap_agent.key",
   "ca_cert" => "/var/vcap/jobs/pcap-agent/config/certs/client_ca.crt",
