+++ jobs/pcap-api/templates/pcap-api.crt.erb
@@ -1,5 +1,5 @@
 <%
-if_p("pcap-api.cert") do |pem|
+if_p("pcap-api.tls.certificate") do |pem|
 %>
 <%= pem %>
 <%
