+++ jobs/pcap-api/templates/pcap-api.key.erb
@@ -1,5 +1,5 @@
 <%
-if_p("pcap-api.key") do |pem|
+if_p("pcap-api.tls.private_key") do |pem|
 %>
 <%= pem %>
 <%
