+++ src/pcap/cmd/pcap-agent-cli/main.go
@@ -17,21 +17,18 @@ package main
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"io"
+	"github.com/cloudfoundry/pcap-release/src/pcap/cmd"
 	"time"
 
 	"github.com/cloudfoundry/pcap-release/src/pcap"
 	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
-	"google.golang.org/grpc/status"
 )
 
 func main() {
 	cc, err := grpc.Dial("localhost:8083", grpc.WithTransportCredentials(insecure.NewCredentials()))
-	p(err)
+	cmd.P(err)
 	agentClient := pcap.NewAgentClient(cc)
 
 	ctx := context.Background()
@@ -39,14 +36,14 @@ func main() {
 	defer cancel()
 
 	statusRes, err := agentClient.Status(ctx, &pcap.StatusRequest{})
-	p(err)
+	cmd.P(err)
 	fmt.Println("status:")
 	fmt.Printf("  healthy: %v\n", statusRes.Healthy)
 	fmt.Printf("  compLvl: %d\n", statusRes.CompatibilityLevel)
 	fmt.Printf("  message: %s\n", statusRes.Message)
 
 	stream, err := agentClient.Capture(ctx)
-	p(err)
+	cmd.P(err)
 
 	err = stream.Send(&pcap.AgentRequest{
 		Payload: &pcap.AgentRequest_Start{
@@ -59,42 +56,14 @@ func main() {
 			},
 		},
 	})
-	p(err)
+	cmd.P(err)
 
-	readN(10, stream)
+	cmd.ReadN(10, stream)
 
 	err = stream.Send(&pcap.AgentRequest{
 		Payload: &pcap.AgentRequest_Stop{},
 	})
-	p(err)
+	cmd.P(err)
 
-	readN(10_000, stream)
-}
-
-func readN(n int, stream pcap.Agent_CaptureClient) {
-	for i := 0; i < n; i++ {
-		res, err := stream.Recv()
-		if errors.Is(err, io.EOF) {
-			fmt.Println("clean stop, done")
-			return
-		}
-		code := status.Code(err)
-		if code != codes.OK {
-			fmt.Printf("receive non-OK code: %s: %s\n", code.String(), err.Error())
-			return
-		}
-
-		switch p := res.Payload.(type) {
-		case *pcap.CaptureResponse_Message:
-			fmt.Printf("received message (%d/%d): %s: %s\n", i+1, n, p.Message.Type.String(), p.Message.Message)
-		case *pcap.CaptureResponse_Packet:
-			fmt.Printf("received packet  (%d/%d): %d bytes\n", i+1, n, len(p.Packet.Data))
-		}
-	}
-}
-
-func p(err error) {
-	if err != nil {
-		panic(err.Error())
-	}
+	cmd.ReadN(10_000, stream)
 }
